requirements.txt
tornado
mysql-connector-python
SQLAlchemy
sockjs-tornado
psutil
pyecharts
pyecharts-snapshot